+++ ext/PlotlyBaseExtensionModule.jl
@@ -2,6 +2,5 @@ module PlotlyBaseExtensionModule
     import PlotlyBase
     import PlotlyExtensionsHelper
 
-    # Load the function for PlotlyBase
-    PlotlyExtensionsHelper._plot_func(::Val{:PlotlyBase}) = PlotlyBase.Plot
+    __init__() = PlotlyExtensionsHelper.register_plot_func!(:PlotlyBase, PlotlyBase.Plot; priority = 0)
 end